refactored:models, serializers & tests
