+++ push/serializers/notification_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import Notification
+
+
+class NotificationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Notification
+        fields = ['id', 'title', 'message', 'action_link', 'status']
