+++ push/serializers/subscription_serializer.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from .models import Subscription, Notification
+from ..models import Subscription
 
 
 # TODO: Will be making the serializer according to the subscription object in the future
@@ -8,8 +8,3 @@ class SubscriptionSerializer(serializers.ModelSerializer):
         model = Subscription
         fields = ['endpoint', 'public_key', 'auth_key']
 
-
-class NotificationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Notification
-        fields = ['id', 'title', 'message', 'action_link', 'status']
