+++ push/tests/__init__.py
@@ -1,0 +1,2 @@
+from .send_notification_api_test import SendNotificationApiTest
+from .subscription_api_test import SubscriptionApiTest
